test: abstract TestUser class from firefuel_collection_test
