+++ test/src/firefuel_collection_test.dart
@@ -3,6 +3,8 @@ import 'package:fake_cloud_firestore/fake_cloud_firestore.dart';
 import 'package:firefuel/firefuel.dart';
 import 'package:test/test.dart';
 
+import '../utils/test_user.dart';
+
 const _testUsersCollectionName = 'testUsers';
 
 void main() {
@@ -286,26 +288,4 @@ class TestCollection extends FirefuelCollection<TestUser> {
 
   @override
   Stream<List<TestUser>> get stream => listenAll(collectionRef);
-}
-
-class TestUser extends Serializable {
-  static const String fieldName = 'name';
-
-  TestUser(this.name, [this.docId]);
-
-  final String name;
-
-  final String? docId;
-
-  @override
-  List<Object?> get props => [name];
-
-  factory TestUser.fromJson(Map<String, dynamic> json, String id) {
-    return TestUser(json[fieldName], id);
-  }
-
-  @override
-  Map<String, dynamic> toJson() {
-    return {fieldName: name};
-  }
 }
