+++ test/utils/test_user.dart
@@ -1,0 +1,23 @@
+import 'package:firefuel/firefuel.dart';
+
+class TestUser extends Serializable {
+  static const String fieldName = 'name';
+
+  TestUser(this.name, [this.docId]);
+
+  final String name;
+
+  final String? docId;
+
+  @override
+  List<Object?> get props => [name];
+
+  factory TestUser.fromJson(Map<String, dynamic> json, String id) {
+    return TestUser(json[fieldName], id);
+  }
+
+  @override
+  Map<String, dynamic> toJson() {
+    return {fieldName: name};
+  }
+}
